+++ packages/sst/src/cli/colors.ts
@@ -1,0 +1,8 @@
+import chalk from "chalk";
+
+export const Colors = {
+  primary: chalk.hex("#E27152"),
+  link: chalk.cyan,
+  success: chalk.green,
+  danger: chalk.red,
+};