+++ packages/sst/src/cli/commands/remove.tsx
@@ -5,26 +5,31 @@ export const remove = (program: Program) =>
     "remove [filter]",
     "Remove your app from AWS",
     (yargs) =>
-      yargs
-        .option("from", { type: "string" })
-        .option("fullscreen", {
-          type: "boolean",
-          describe: "Disable full screen UI",
-          default: true,
-        })
-        .positional("filter", {
-          type: "string",
-          describe: "Optionally filter stacks to remove",
-        }),
+      yargs.option("from", { type: "string" }).positional("filter", {
+        type: "string",
+        describe: "Optionally filter stacks to remove",
+      }),
     async (args) => {
       const React = await import("react");
-      const { blue, bold } = await import("colorette");
+      const { dim, blue, bold } = await import("colorette");
       const { useProject } = await import("../../project.js");
       const { loadAssembly, Stacks } = await import("../../stacks/index.js");
       const { render } = await import("ink");
       const { DeploymentUI } = await import("../ui/deploy.js");
       const { printDeploymentResults } = await import("../ui/deploy.js");
-      const { createSpinner } = await import("../spinner.js");
+      const { Colors } = await import("../colors.js");
+
+      const project = useProject();
+
+      console.log();
+      console.log(`  ${Colors.primary(`${bold(`SST`)} v${project.version}`)}`);
+      console.log();
+      console.log(
+        `  ${Colors.primary(`➜`)}  ${bold(`Stage:`)}   ${dim(
+          project.config.stage
+        )}`
+      );
+      console.log();
 
       const assembly = await (async function () {
         if (args.from) {
@@ -32,14 +37,12 @@ export const remove = (program: Program) =>
           return result;
         }
 
-        const project = useProject();
         return await Stacks.synth({
           fn: project.stacks,
           mode: "remove",
         });
       })();
 
-      const project = useProject();
       const target = assembly.stacks.filter(
         (s) =>
           !args.filter ||
@@ -49,29 +52,13 @@ export const remove = (program: Program) =>
         console.log(`No stacks found matching ${blue(args.filter!)}`);
         process.exit(1);
       }
-      console.log(
-        `Removing ${bold(target.length + " stacks")} for stage ${blue(
-          project.config.stage
-        )}...`
+      const component = render(
+        <DeploymentUI stacks={assembly.stacks.map((s) => s.stackName)} />
       );
-      const cleanup = (() => {
-        if (args.fullscreen) {
-          process.stdout.write("\x1b[?1049h");
-          const component = render(
-            <DeploymentUI stacks={assembly.stacks.map((s) => s.stackName)} />
-          );
-          return () => {
-            component.unmount();
-            process.stdout.write("\x1b[?1049l");
-          };
-        }
-
-        const spinner = createSpinner("Removing stacks");
-        return () => spinner.succeed();
-      })();
       const results = await Stacks.removeMany(target);
-      cleanup();
-      printDeploymentResults(results);
+      component.clear();
+      component.unmount();
+      printDeploymentResults(assembly, results);
       if (Object.values(results).some((stack) => Stacks.isFailed(stack.status)))
         process.exit(1);
       process.exit(0);