+++ packages/sst/src/cli/ui/deploy.tsx
@@ -1,41 +1,16 @@
 import React, { useState, useEffect } from "react";
-import type { StackResource } from "@aws-sdk/client-cloudformation";
+import type { StackEvent, StackResource } from "@aws-sdk/client-cloudformation";
 import { Box, Text } from "ink";
 import { useBus } from "../../bus.js";
 import { Stacks } from "../../stacks/index.js";
 import inkSpinner from "ink-spinner";
 import { useProject } from "../../project.js";
-import { blue, bold, green, red } from "colorette";
+import { blue, bold, dim, green, red } from "colorette";
+import { Colors } from "../colors.js";
+import { CloudAssembly } from "aws-cdk-lib/cx-api";
 
 // @ts-ignore
 const { default: Spinner } = inkSpinner;
-
-const FullScreen: React.FC = (props) => {
-  const [size, setSize] = useState({
-    columns: process.stdout.columns,
-    rows: process.stdout.rows,
-  });
-
-  useEffect(() => {
-    function onResize() {
-      setSize({
-        columns: process.stdout.columns,
-        rows: process.stdout.rows,
-      });
-    }
-
-    process.stdout.on("resize", onResize);
-    return () => {
-      process.stdout.off("resize", onResize);
-    };
-  }, []);
-
-  return (
-    <Box width={size.columns} height={size.rows} flexDirection="column">
-      {props.children}
-    </Box>
-  );
-};
 
 interface Props {
   stacks: string[];
@@ -49,7 +24,10 @@ export const DeploymentUI = (props: Props) => {
     {}
   );
 
+  const [resources2, setResources2] = useState<Record<string, StackEvent>>({});
+
   useEffect(() => {
+    console.log(`  ${Colors.primary(`➜`)}  ${bold(dim(`Deploying...`))}`);
     const bus = useBus();
 
     const update = bus.subscribe("stack.updated", (payload) => {
@@ -66,24 +44,33 @@ export const DeploymentUI = (props: Props) => {
       }));
     });
 
-    const resourcesSub = bus.subscribe("stack.resources", (payload) => {
-      setResources((previous) => {
-        const existing = previous[payload.properties.stackID] || [];
-        const filtered = (payload.properties.resources || []).filter((r) => {
-          if (existing.some((p) => p.LogicalResourceId === r.LogicalResourceId))
-            return true;
-          if (!Stacks.isFinal(r.ResourceStatus as any)) return true;
-          return false;
-        });
+    const event = bus.subscribe("stack.event", (payload) => {
+      const { event } = payload.properties;
+      if (event.ResourceType === "AWS::CloudFormation::Stack") return;
+      setResources2((previous) => {
+        if (Stacks.isFinal(event.ResourceStatus!)) {
+          console.log(
+            dim(
+              `     ${event.StackName} ${event.ResourceType} ${event.LogicalResourceId}`
+            ),
+            Stacks.isFailed(event.ResourceStatus!)
+              ? Colors.danger(event.ResourceStatus!)
+              : dim(event.ResourceStatus!)
+          );
+          const { [event.LogicalResourceId!]: _, ...next } = previous;
+          return next;
+        }
+
         return {
           ...previous,
-          [payload.properties.stackID]: filtered,
+          [payload.properties.event.LogicalResourceId!]:
+            payload.properties.event,
         };
       });
     });
 
     return () => {
-      bus.unsubscribe(resourcesSub);
+      bus.unsubscribe(event);
       bus.unsubscribe(update);
       bus.unsubscribe(status);
     };
@@ -96,84 +83,84 @@ export const DeploymentUI = (props: Props) => {
   }
 
   return (
-    <FullScreen>
-      <Text>
-        Deploying <Text color="bold">{props.stacks.length}</Text> stack
-        {props.stacks.length > 1 && "s"} for stage{" "}
-        <Text color="blue">{useProject().config.stage}</Text>
-      </Text>
-      {Object.entries(stacks).map(([stackID, status]) => {
+    <Box flexDirection="column">
+      {Object.entries(resources2).map(([_, evt]) => {
         return (
-          <React.Fragment key={stackID}>
+          <Box key={evt.LogicalResourceId}>
             <Text>
-              {!Stacks.isFinal(status) && <Spinner />}
-              {Stacks.isSuccess(status) && <Text color={color(status)}>✔</Text>}
-              {Stacks.isFailed(status) && <Text color={color(status)}>✖</Text>}
-              <Text>{" " + stackID}</Text>
-              {status && <Text color={color(status)}> {status}</Text>}
+              {"  "}
+              <Spinner />
+              {"  "}
+              {evt.StackName} {evt.ResourceType} {evt.LogicalResourceId}{" "}
             </Text>
-            {resources[stackID]?.map((resource) => (
-              <Box key={resource.LogicalResourceId}>
-                <Text>
-                  {"  "}
-                  {!Stacks.isFinal(resource.ResourceStatus || "") && (
-                    <Spinner />
-                  )}
-                  {Stacks.isSuccess(resource.ResourceStatus || "") && (
-                    <Text color="green">✔</Text>
-                  )}
-                  {Stacks.isFailed(resource.ResourceStatus || "") && (
-                    <Text color="red">✖</Text>
-                  )}{" "}
-                  {resource.ResourceType} {resource.LogicalResourceId}{" "}
-                  {resource.ResourceStatusReason}{" "}
-                </Text>
-                <Text color={color(resource.ResourceStatus || "")}>
-                  {resource.ResourceStatus}
-                </Text>
-              </Box>
-            ))}
-          </React.Fragment>
+            <Text color={color(evt.ResourceStatus || "")}>
+              {evt.ResourceStatus}
+            </Text>
+          </Box>
         );
       })}
-    </FullScreen>
+      {/*
+      <Box>
+        <Text>
+          {"     "}
+          <Spinner />
+          <Text dimColor> Polling for updates </Text>
+        </Text>
+      </Box>
+      */}
+      <Box />
+    </Box>
   );
 };
 
 export function printDeploymentResults(
+  assembly: CloudAssembly,
   results: Awaited<ReturnType<typeof Stacks.deployMany>>
 ) {
   console.log();
-  console.log(`-----------`);
-  console.log(`| Summary |`);
-  console.log(`-----------`);
-  for (const [stack, result] of Object.entries(results)) {
-    const icon = (() => {
-      if (Stacks.isSuccess(result.status)) return green("✔");
-      if (Stacks.isFailed(result.status)) return red("✖");
-    })();
-    console.log(`${icon} ${stack}`);
 
+  console.log(`  ${green(`✔`)}  ${bold(dim(`Deployed:`))}`);
+  for (const [stack, result] of Object.entries(results)) {
     const outputs = Object.entries(result.outputs).filter(([key, _]) => {
       if (key.startsWith("Export")) return false;
       if (key.includes("SstSiteEnv")) return false;
       if (key === "SstMetadata") return false;
       return true;
     });
+    console.log(`     ${dim(stack)}`);
     if (outputs.length > 0) {
-      console.log(`  ${blue("Outputs:")}`);
       for (const key of Object.keys(Object.fromEntries(outputs)).sort()) {
         const value = result.outputs[key];
-        console.log(bold(`    ${key}: ${value}`));
-      }
-    }
-
-    if (Object.entries(result.errors).length > 0) {
-      console.log(`  ${red("Errors:")}`);
-      for (const [id, error] of Object.entries(result.errors)) {
-        console.log(bold(`    ${id}: ${error}`));
+        console.log(`     ${bold(dim(key))}: ${value}`);
       }
     }
   }
   console.log();
+
+  if (Object.values(results).flatMap((s) => Object.keys(s.errors)).length) {
+    console.log(`  ${red(`✖`)}  ${bold(dim(`Errors:`))}`);
+    for (const [stack, result] of Object.entries(results)) {
+      const hasErrors = Object.entries(result.errors).length > 0;
+      if (!hasErrors) continue;
+      console.log(`     ${dim(stack)}`);
+      for (const [id, error] of Object.entries(result.errors)) {
+        const found =
+          Object.entries(
+            assembly.manifest.artifacts?.[stack].metadata || {}
+          ).find(
+            ([_key, value]) =>
+              value[0]?.type === "aws:cdk:logicalId" && value[0]?.data === id
+          )?.[0] || "";
+
+        const readable = found
+          .split("/")
+          .filter(Boolean)
+          .slice(1, -1)
+          .join("/");
+
+        console.log(`     ${bold(red(readable))}: ${error}`);
+      }
+      console.log();
+    }
+  }
 }