+++ packages/sst/support/bridge/bridge.ts
@@ -94,6 +94,7 @@ export async function handler(event: any, context: any) {
     type: "function.invoked",
     properties: {
       workerID: workerID,
+      requestID: context.awsRequestId,
       functionID: process.env.SST_FUNCTION_ID,
       deadline: context.getRemainingTimeInMillis(),
       event,
@@ -112,11 +113,15 @@ export async function handler(event: any, context: any) {
     }, 5 * 1000);
     onMessage = (evt) => {
       if (evt.type === "function.ack") {
-        if (evt.properties.workerID === workerID) return;
-        clearTimeout(timeout);
+        if (evt.properties.workerID === workerID) {
+          clearTimeout(timeout);
+        }
       }
       if (["function.success", "function.error"].includes(evt.type)) {
-        if (evt.properties.workerID === workerID) r(evt);
+        if (evt.properties.workerID === workerID) {
+          clearTimeout(timeout);
+          r(evt);
+        }
       }
     };
   });